Move KinitAuxiliaryService to kyuubi-server module
